+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Dtos/Live/HeartBeatRequest.cs
@@ -8,7 +8,7 @@ namespace Ray.BiliBiliTool.Agent.BiliBiliAgent.Dtos.Live
     public class HeartBeatRequest
     {
         public HeartBeatRequest(
-            int roomId,
+            long roomId,
             int parentId,
             int areaID,
             int seqNumber, // 心跳包编号
