+++ src/Ray.BiliBiliTool.DomainService/Dtos/FansMedalInfoDto.cs
@@ -9,7 +9,7 @@ namespace Ray.BiliBiliTool.DomainService.Dtos
 {
     public class FansMedalInfoDto
     {
-        public FansMedalInfoDto(int roomId, MedalWallDto medalInfo, GetLiveRoomInfoResponse liveRoomInfo)
+        public FansMedalInfoDto(long roomId, MedalWallDto medalInfo, GetLiveRoomInfoResponse liveRoomInfo)
         {
             this.RoomId = roomId;
             this.MedalInfo = medalInfo;
@@ -17,7 +17,7 @@ namespace Ray.BiliBiliTool.DomainService.Dtos
         }
 
         // 直播间 id
-        public int RoomId { get; set; }
+        public long RoomId { get; set; }
 
         // 粉丝牌信息
         public MedalWallDto MedalInfo { get; set; }
